# src/main/java/com/example/deviceapi/dto/DeviceResponse.java
package com.example.deviceapi.dto;

import com.example.deviceapi.model.DeviceState;
import java.time.LocalDateTime;

/**
 * DTO representing the response payload for Device API operations.
 * 
 * This immutable record encapsulates the device's core details returned to the
 * client after creation, retrieval, or update.
 * 
 * Fields include: - id: unique identifier of the device (generated by the
 * system) - name: device name or model designation - brand: device manufacturer
 * or brand name - state: current device status (enum) - creationTime: timestamp
 * when the device was registered in the system
 * 
 * Using a record ensures immutability and thread safety, preventing accidental
 * modification of response data.
 */
public record DeviceResponse(Long id, String name, String brand, DeviceState state, LocalDateTime creationTime) {
}